+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.1.0 (2019-03-19)
+------------------
+
+* Ensure that inline static included the hashed versions of a static file when
+  filename hashing is available in the staticfiles storage.
+
+
 0.0.1 (2018-08-08)
 ------------------